Update variable types to long long in H-GCD.cpp
Changed variable types from int to long long for larger input handling.

+++ sheet6/H-GCD.cpp
@@ -7,10 +7,10 @@ int main()
     ios_base::sync_with_stdio(0);
     cin.tie(0);
     cout.tie(0);
-    int a, b;
+   long long a, b;
     cin >> a >> b;
-    cout << gcd(a, b)<<" ";
+    cout << gcd(a, b) << " ";
     cout << lcm(a, b);
     return 0;
-
 }
+
